Rework the samples to have framework/core/common
