+++ samples/ConsoleSample.AspNetCore/Program.cs
@@ -1,0 +1,12 @@
+﻿using System;
+
+namespace ConsoleSample.AspNetCore
+{
+    internal class Program
+    {
+        static void Main(string[] args)
+        {
+            Console.WriteLine("Hello World!");
+        }
+    }
+}
