--- samples/ConsoleSample/ConsoleSample.csproj
@@ -1,11 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <OutputType>Exe</OutputType>
-    <TargetFramework>net6.0</TargetFramework>
-    <ImplicitUsings>enable</ImplicitUsings>
-    <Nullable>disable</Nullable>
-    <WarningsAsErrors>true</WarningsAsErrors>
-  </PropertyGroup>
-
-</Project>
