--- samples/ConsoleSample/FeaturesManagement/Features.cs
@@ -1,24 +1,0 @@
-using System.ComponentModel;
-using System.Reflection;
-
-namespace ConsoleSample.FeaturesManagement;
-
-public static class Features
-{
-    // avoid feature flag names with spaces, Pascal-case is good, sticking to just letters/numbers is good.
-    
-    [Description("A: some longer explanation goes here")]
-    public const string A = nameof(A);
-
-    [Description("Fancy New Interface: New and improved! Now with Brawndo!")]
-    public const string FancyNewInterface = nameof(FancyNewInterface);
-
-    public static readonly Lazy<IEnumerable<string>> All = new(() =>
-    {
-        return typeof(Features)
-            .GetFields(BindingFlags.Public | BindingFlags.Static | BindingFlags.FlattenHierarchy)
-            .Where(fi => fi.IsLiteral && !fi.IsInitOnly && fi.FieldType == typeof(string))
-            .Select(x => (string)x.GetRawConstantValue())
-            .ToList();
-    });
-}
